# mutex: Improve performance of valid function