The `valid()` function of `mutex_array` tests each bit individually and performs a reduction to check if all mutexes are unlocked. Use the `count()` function of `bitset` instead to simplify the code. This also significantly improves the performance for large container sizes due to the more efficient reduction of the bits in `bitset`.

Before:
```
----------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations
----------------------------------------------------------------------
stdgpu_mutex_valid/1000          0.027 ms        0.027 ms        25450
stdgpu_mutex_valid/100000        0.051 ms        0.051 ms        12299
stdgpu_mutex_valid/10000000      0.695 ms        0.694 ms          991
```

After:
```
----------------------------------------------------------------------
Benchmark                            Time             CPU   Iterations
----------------------------------------------------------------------
stdgpu_mutex_valid/1000          0.049 ms        0.049 ms        13793
stdgpu_mutex_valid/100000        0.050 ms        0.050 ms        13565
stdgpu_mutex_valid/10000000      0.302 ms        0.301 ms         2316
```